% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf-api.r
\name{add_key}
\alias{add_key}
\title{Add MySportsFeeds API key to R environment file}
\usage{
add_key(key)
}
\arguments{
\item{key}{API key provided by MySportsFeeds}
}
\description{
The functions in this package will look for your MySportsFeeds API key
in a special environment variable \code{MSF}. This function opens the .Renviron
file and tells what line you should add. Just copy-and-paste the line into the file,
save it, and restart R.
}
