% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{delay_by}
\alias{delay_by}
\title{Delays function result}
\usage{
delay_by(delay, f)
}
\arguments{
\item{delay}{number of seconds to delay function by}

\item{f}{function}
}
\description{
Delays function result
}
\keyword{internal}
