% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf-by-date.r
\name{msf_by_date}
\alias{msf_by_date}
\alias{daily_dfs}
\alias{daily_games}
\title{MySportsFeeds API calls by date}
\usage{
msf_by_date(feed, sport, date = Sys.Date(), season = "current",
  delay = 1)

daily_dfs(sport, ...)

daily_games(sport, ...)
}
\arguments{
\item{feed}{feed to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}

\item{delay}{number of seconds between queries}

\item{...}{additional parameters passed to \code{msf_by_date} such as date or season}
}
\description{
If you provide multiple dates it will execute with a delay between each query.
This is to ensure you obey limit rates but you are free to lower the parameter if you don't
think your queries will reach the throttle limits.
}
\section{Functions}{
\itemize{
\item \code{daily_dfs}: A list of players, along with their DFS salaries and actual fantasy points.

\item \code{daily_games}: All games on a given date including schedule, status and scores.
}}

\examples{
\dontrun{
j <- daily_dfs("mlb", "2018-regular", date = as.Date("2018-07-15"))
}
}
