% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf-by-game.r
\name{msf_by_game}
\alias{msf_by_game}
\alias{game_lineup}
\alias{game_boxscore}
\alias{game_pbp}
\title{MySportsFeeds API calls by game}
\usage{
msf_by_game(feed, sport, gameid, season = "current", delay = 1, ...)

game_lineup(sport, gameid, season = "current", ...)

game_boxscore(sport, gameid, season = "current", ...)

game_pbp(sport, gameid, season = "current", ...)
}
\arguments{
\item{feed}{feed to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{strings, "40265" or c("40265", "11111") or "20171015-SF-WAS"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}

\item{delay}{number of seconds between queries}

\item{...}{additional query parameters}
}
\description{
If you provide multiple gameids it will execute with a delay between each query.
This is to ensure you obey limit rates but you are free to lower the parameter if you don't
think your queries will reach the throttle limits.
}
\section{Functions}{
\itemize{
\item \code{game_lineup}: Expected and Actual starting lineup for both teams of a game.

\item \code{game_boxscore}: A boxscore for a game including game details and team/player stats.

\item \code{game_pbp}: All play-by-play events for a game with full details.
}}

\examples{
\dontrun{
j <- game_boxscore("nba", "47627", playerstats = c("ftMade", "ftAtt"))
}
}
