% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf-by-season.r
\name{msf_by_season}
\alias{msf_by_season}
\alias{seasonal_dfs}
\title{MySportsFeeds API calls by season}
\usage{
msf_by_season(feed, sport, season = "current", delay = 1, ...)

seasonal_dfs(sport, season, ...)
}
\arguments{
\item{feed}{feed to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}

\item{delay}{number of seconds between queries}

\item{...}{additional query parameters}
}
\description{
If you provide multiple seasons it will execute with a delay between each query.
This is to ensure you obey limit rates but you are free to lower the parameter if you don't
think your queries will reach the throttle limits.
}
\section{Functions}{
\itemize{
\item \code{seasonal_dfs}: All DFS entries for a season including salaries and fantasy points.
}}

\examples{
\dontrun{
j <- seasonal_dfs("nba", "2017-2018-regular", team = "SAS")
}
}
