% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{msf_class}
\alias{msf_class}
\title{Adds msf_* type to object class}
\usage{
msf_class(obj, newclass)
}
\arguments{
\item{obj}{object to modify}

\item{newclass}{new class will be msf_newclass}
}
\description{
Adds msf_* type to object class
}
\keyword{internal}
