% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{msf_date}
\alias{msf_date}
\title{Converts a Date to msf formatted date}
\usage{
msf_date(date)
}
\arguments{
\item{date}{an R date}
}
\description{
MSF dates have a specific format. This will parse R dates to MSF format.
}
\keyword{internal}
