% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{msf_time}
\alias{msf_time}
\title{Converts msf formatted time to POSIXct}
\usage{
msf_time(x)
}
\arguments{
\item{x}{character string}
}
\description{
MSF datetimes have a specific format. This will parse them to R datetimes.
}
\keyword{internal}
