% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-game.r
\name{tidy.msf_boxscore}
\alias{tidy.msf_boxscore}
\title{Tidy Game Boxscores}
\usage{
\method{tidy}{msf_boxscore}(j, ...)
}
\arguments{
\item{j}{msf object}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy Game Boxscores
}
