% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-date.r
\name{tidy.msf_dfs}
\alias{tidy.msf_dfs}
\title{Tidy DFS
Converts daily and seasonal dfs data to a dataframe}
\usage{
\method{tidy}{msf_dfs}(j, site = c("draftkings", "fanduel"), ...)
}
\arguments{
\item{j}{msf object}

\item{site}{DFS site to get data for}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy DFS
Converts daily and seasonal dfs data to a dataframe
}
