% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-date.r
\name{tidy.msf_games}
\alias{tidy.msf_games}
\title{Tidy Daily Games}
\usage{
\method{tidy}{msf_games}(j, ...)
}
\arguments{
\item{j}{msf object}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy Daily Games
}
