% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-game.r
\name{tidy.msf_lineup}
\alias{tidy.msf_lineup}
\title{Tidy Starting Lineups}
\usage{
\method{tidy}{msf_lineup}(j, ...)
}
\arguments{
\item{j}{msf object}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy Starting Lineups
}
