% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-date.r
\name{tidy.msf_odds_gamelines}
\alias{tidy.msf_odds_gamelines}
\title{Tidy Daily Odds}
\usage{
\method{tidy}{msf_odds_gamelines}(j, ...)
}
\arguments{
\item{j}{msf object}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy Daily Odds
}
