% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-players.r
\name{tidy.msf_players}
\alias{tidy.msf_players}
\title{Tidy Players}
\usage{
\method{tidy}{msf_players}(j, ...)
}
\arguments{
\item{j}{msf object}

\item{...}{additional arguments. currently unused}
}
\description{
Tidy Players
}
