% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-msf-by-date.r
\name{tidy_lines}
\alias{tidy_lines}
\title{Tidy MoneyLines}
\usage{
tidy_lines(j)
}
\arguments{
\item{j}{json data}
}
\description{
Currently only exracts moneylines, not point spreads or over/unders.
}
\keyword{internal}
