% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.r
\name{write_msf}
\alias{write_msf}
\title{Write msf api objects to disk}
\usage{
write_msf(j, datadir = ".")
}
\arguments{
\item{j}{a list representing a json object}

\item{datadir}{base directory to write files into}
}
\description{
Write msf api objects to disk
}
