% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf-by-game.r
\name{msf_by_game}
\alias{msf_by_game}
\alias{game_lineup}
\title{MySportsFeeds API calls by game}
\usage{
msf_by_game(feed, sport, gameid, season = "current", delay = 1)

game_lineup(sport, gameid, ...)
}
\arguments{
\item{feed}{feed to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{strings, "40265" or c("40265", "11111") or "20171015-SF-WAS"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}

\item{delay}{number of seconds between queries}

\item{...}{additional parameters passed to \code{msf_by_game} such as season or delay}
}
\description{
If you provide multiple gameids it will execute with a delay between each query.
This is to ensure you obey limit rates but you are free to lower the parameter if you don't
think your queries will reach the throttle limits.
}
\section{Functions}{
\itemize{
\item \code{game_lineup}: Expected and Actual starting lineup for both teams of a game.
}}

\examples{
\dontrun{
j <- game_lineup("mlb", "43333", season = "2017-regular")
}
}
